chore: move axis fitting into shared
Both sides fit axes: the server for the window it queried, and the client
again when a zoom narrows that window to a sub-range it already holds. Two
implementations would drift.
