+++ shared/src/chart/tests.rs
@@ -376,3 +376,72 @@ fn every_measure_carries_a_sentence_saying_what_it_counts() {
         assert_ne!(d.trim_end_matches('.'), m.label());
     }
 }
+
+// ── Axis fitting ────────────────────────────────────────────────────────
+
+#[test]
+fn nice_step_rounds_up_to_a_readable_tick() {
+    assert_eq!(nice_step(0.9), 1.0);
+    assert_eq!(nice_step(1.2), 2.0);
+    assert_eq!(nice_step(2.4), 2.5);
+    assert_eq!(nice_step(137.5), 200.0);
+}
+
+#[test]
+fn nice_axis_keeps_the_data_filling_most_of_the_frame() {
+    // The case the old round-the-maximum ladder got wrong: a peak of 550
+    // took a 1000 axis and drew every shape at half height.
+    let (max, divisions) = nice_axis(550.0, false);
+    assert_eq!((max, divisions), (600.0, 3));
+    assert!(550.0 / max > 0.9);
+
+    // Stars: a 4.5 peak must not open an axis past the 5 they top out at.
+    assert_eq!(nice_axis(4.5, false), (5.0, 5));
+}
+
+#[test]
+fn nice_axis_gives_a_small_count_one_gridline_per_whole_unit() {
+    // Half a book is not a quantity an axis should offer to measure.
+    assert_eq!(nice_axis(6.0, true), (6.0, 6));
+    assert_eq!(nice_axis(2.0, true), (2.0, 2));
+    assert_eq!(nice_axis(1.0, true), (1.0, 1));
+    // Without the integral flag the same peak takes the ladder's answer.
+    assert_eq!(nice_axis(2.0, false), (2.0, 4));
+}
+
+#[test]
+fn nice_axis_keeps_a_large_count_on_whole_number_steps() {
+    let (max, divisions) = nice_axis(38.0, true);
+    assert_eq!((max, divisions), (40.0, 4));
+    // Every gridline lands on an integer.
+    assert_eq!(max % divisions as f64, 0.0);
+}
+
+#[test]
+fn nice_axis_never_returns_a_max_below_the_data() {
+    for peak in [0.3, 1.0, 6.0, 47.0, 317.0, 550.0, 999.0, 1_000_000.0] {
+        for integral in [false, true] {
+            let (max, _) = nice_axis(peak, integral);
+            assert!(max >= peak, "axis {max} is under its peak {peak}");
+        }
+    }
+}
+
+#[test]
+fn nice_axis_never_returns_zero_so_the_renderer_cannot_divide_by_it() {
+    assert_eq!(nice_axis(0.0, false).0, 1.0);
+    assert_eq!(nice_axis(-5.0, false).0, 1.0);
+    assert_eq!(nice_axis(f64::NAN, false).0, 1.0);
+    assert_eq!(nice_axis(0.0, true).0, 1.0);
+    assert_eq!(axis_max_on(0.0, 4, false), 1.0);
+}
+
+#[test]
+fn both_axes_share_one_gridline_count() {
+    // The right axis is fitted to the left's tick count, so a label on the
+    // right always sits on a line that is actually drawn.
+    let divisions = 5;
+    let right = axis_max_on(450.0, divisions, false);
+    assert_eq!(right, 500.0);
+    assert!((right / divisions as f64 - 100.0).abs() < f64::EPSILON);
+}
